app/requirements.txt
fastapi
pydantic
python-dotenv
openai
uvicorn